+++ app/assets/stylesheets/_footer.scss
@@ -5,7 +5,6 @@
 .footer {
   display: block;
   width: 100%;
-  margin-top: 2em;
   text-align: center;
 
   h5, p , a, a:hover {
