+++ app/assets/stylesheets/new/_variables.scss
@@ -12,3 +12,4 @@ $sponge_dark_grey: #2b2b2b;
 $neon-red: #ff544b;
 $neon-yellow: #f0d95c;
 $neon-green: #3fa763;
+$neon-blue: #5fa9bd;
