+++ app/controllers/Application.scala
@@ -167,90 +167,6 @@ final class Application @Inject()(forms: OreForms)(
   }
 
   /**
-    * Shows the moderation queue for unreviewed versions.
-    *
-    * @return View of unreviewed versions.
-    */
-  def showVersionQueuePending(page: Option[Int]): Action[AnyContent] = (Authenticated andThen PermissionAction[AuthRequest](ReviewProjects)).async { implicit request =>
-    service.DB.db.run(queryVersionQueuePending(page.getOrElse(1)).result).map { list: Seq[(Version, Project, Channel, Option[String], User, Option[Review])] =>
-      Ok(views.users.admin.queue.pending(list))
-    }
-  }
-
-  def showVersionQueueWIP(page: Option[Int]): Action[AnyContent] = (Authenticated andThen PermissionAction[AuthRequest](ReviewProjects)).async { implicit request =>
-    service.DB.db.run(queryVersionQueueWIP(page.getOrElse(1)).result).map { list: Seq[(Version, Project, Channel, Option[String], User, Review)] =>
-      Ok(views.users.admin.queue.wip(list))
-    }
-  }
-
-    def showVersionQueueDone(page: Option[Int]): Action[AnyContent] = (Authenticated andThen PermissionAction[AuthRequest](ReviewProjects)).async { implicit request =>
-    service.DB.db.run(queryVersionQueueDone(page.getOrElse(1)).result).map { list: Seq[(Version, Project, Channel, Option[String], User)] =>
-      Ok(views.users.admin.queue.done(list))
-    }
-    }
-
-  private def queuePageSize = this.config.ore.get[ObjectReference]("queue.page-size")
-
-  private def queryVersionQueuePending(page: Int = 1) = {
-    val versionTable = TableQuery[VersionTable]
-    val channelTable = TableQuery[ChannelTable]
-    val projectTable = TableQuery[ProjectTableMain]
-    val userTable = TableQuery[UserTable]
-    val reviewsTable = TableQuery[ReviewTable]
-
-    (for {
-      ((v, u), r) <- (versionTable joinLeft userTable on (_.authorId === _.id)) joinLeft reviewsTable on (_._1.id === _.versionId)
-      c <- channelTable if v.channelId === c.id && v.isReviewed =!= true && v.isNonReviewed =!= true && v.isReviewed === false && v.visibility =!= VisibilityTypes.SoftDelete
-      p <- projectTable if v.projectId === p.id && p.visibility =!= VisibilityTypes.SoftDelete
-      ou <- userTable if p.userId === ou.id
-    } yield {
-      (v, p, c, u.map(_.name), ou, r)
-    }).filter { case (_, _, _, _, _, r) =>
-        r.isEmpty
-    }.sortBy { case (v, _, _, _, _, _) =>
-      v.createdAt.asc.nullsFirst
-    }.drop((page - 1) * queuePageSize).take(queuePageSize)
-  }
-
-  private def queryVersionQueueWIP(page: Int = 1) = {
-    val versionTable = TableQuery[VersionTable]
-    val channelTable = TableQuery[ChannelTable]
-    val projectTable = TableQuery[ProjectTableMain]
-    val userTable = TableQuery[UserTable]
-    val reviewsTable = TableQuery[ReviewTable]
-
-    (for {
-      (v, u) <- versionTable joinLeft userTable on (_.authorId === _.id)
-      c <- channelTable if v.channelId === c.id && v.isReviewed =!= true && v.isNonReviewed =!= true && v.isReviewed === false && v.visibility =!= VisibilityTypes.SoftDelete
-      p <- projectTable if v.projectId === p.id && p.visibility =!= VisibilityTypes.SoftDelete
-      ou <- userTable if p.userId === ou.id
-      r <- reviewsTable if r.versionId === v.id
-    } yield {
-      (v, p, c, u.map(_.name), ou, r)
-    }).sortBy { case (_, _, _, _, _, r) =>
-      r.createdAt.desc.nullsLast
-    }.drop((page - 1) * queuePageSize).take(queuePageSize)
-  }
-
-  private def queryVersionQueueDone(page: Int = 1) = {
-    val versionTable = TableQuery[VersionTable]
-    val channelTable = TableQuery[ChannelTable]
-    val projectTable = TableQuery[ProjectTableMain]
-    val userTable = TableQuery[UserTable]
-
-    (for {
-      (v, u) <- versionTable joinLeft userTable on (_.authorId === _.id)
-      c <- channelTable if v.channelId === c.id && v.isNonReviewed =!= true && c.isNonReviewed =!= true && v.isReviewed === true && v.visibility =!= VisibilityTypes.SoftDelete
-      p <- projectTable if v.projectId === p.id && p.visibility =!= VisibilityTypes.SoftDelete
-      ou <- userTable if p.userId === ou.id
-    } yield {
-      (v, p, c, u.map(_.name), ou)
-    }).sortBy { case (v, _, _, _, _) =>
-      v.approvedAt.desc.nullsLast
-    }.drop((page - 1) * queuePageSize).take(queuePageSize)
-  }
-
-  /**
     * Shows the overview page for flags.
     *
     * @return Flag overview
