+++ app/views/bootstrap/footer.scala.html
@@ -1,5 +1,5 @@
 @()(implicit messages: Messages)
-<footer class="footer">
+<footer class="footer mt-3">
     <div class="above">
         <div class="container">
             <div class="row">
