+++ app/views/users/admin/queue/template.scala.html
@@ -12,14 +12,86 @@
 @import java.sql.Timestamp
 @import java.time.Instant
 
-@(currentTab: Int)(content: Html)(implicit messages: Messages, request: OreRequest[_], config: OreConfig, flash: Flash)
+@(currentTab: Int, currentPage: Int, totalPages: Int)(content: Html)(implicit messages: Messages, request: OreRequest[_], config: OreConfig, flash: Flash)
+
+@checkMaxMin(page: Int) = @{
+    if(page <= 0) {
+        1
+    } else if (page > totalPages) {
+        totalPages
+    } else {
+        page
+    }
+}
+
+@currentLink(page: Int) = @{
+    var somePage = Some(page)
+
+    if(currentTab == 2) {
+        routes.Reviews.showVersionQueueWIP(somePage)
+    } else if(currentTab == 3) {
+        routes.Reviews.showVersionQueueDone(somePage)
+    } else {
+        routes.Reviews.showVersionQueuePending(somePage)
+    }
+}
 
 @bootstrap.layout(messages("user.queue")) {
     <nav class="nav nav-pills flex-column flex-sm-row mb-3">
-        <a class="flex-sm-fill text-sm-center nav-link @if(currentTab == 1) { active }" href="@routes.Application.showVersionQueuePending(None)">Pending</a>
-        <a class="flex-sm-fill text-sm-center nav-link @if(currentTab == 2) { active }" href="@routes.Application.showVersionQueueWIP(None)">Work in Progess</a>
-        <a class="flex-sm-fill text-sm-center nav-link @if(currentTab == 3) { active }" href="@routes.Application.showVersionQueueDone(None)">Done</a>
+        <a class="flex-sm-fill text-sm-center nav-link @if(currentTab == 1) { active }" href="@routes.Reviews.showVersionQueuePending(None)">Pending</a>
+        <a class="flex-sm-fill text-sm-center nav-link @if(currentTab == 2) { active }" href="@routes.Reviews.showVersionQueueWIP(None)">Work in Progess</a>
+        <a class="flex-sm-fill text-sm-center nav-link @if(currentTab == 3) { active }" href="@routes.Reviews.showVersionQueueDone(None)">Done</a>
     </nav>
 
-    @content
+    <div class="row">
+        <div class="col">
+            @content
+        </div>
+    </div>
+
+    <div class="row">
+        <div class="col">
+            <nav class="mt-3" aria-label="Review page navigation">
+                <ul class="pagination justify-content-center">
+                    <li class="page-item @if(currentPage == 1) { disabled }">
+                        <a class="page-link" href="@currentLink(checkMaxMin(currentPage - 1))" aria-label="Previous">
+                            <span aria-hidden="true">&laquo;</span>
+                            <span class="sr-only">Previous</span>
+                        </a>
+                    </li>
+
+                    @if(currentPage > 2) {
+                        <li class="page-item"><a class="page-link" href="@currentLink(1)">1</a></li>
+                    }
+                    @if(currentPage > 3) {
+                        <li class="page-item disabled"><a class="page-link">...</a></li>
+                    }
+
+                    @if(currentPage > 1) {
+                        <li class="page-item"><a class="page-link" href="@currentLink(currentPage - 1)">@(currentPage - 1)</a></li>
+                    }
+
+                    <li class="page-item active"><a class="page-link">@currentPage</a></li>
+
+                    @if(currentPage != totalPages) {
+                        <li class="page-item"><a class="page-link" href="@currentLink(currentPage + 1)">@(currentPage + 1)</a></li>
+                    }
+
+                    @if(totalPages - currentPage > 2) {
+                        <li class="page-item disabled"><a class="page-link">...</a></li>
+                    }
+                    @if(totalPages - currentPage > 1) {
+                        <li class="page-item"><a class="page-link" href="@currentLink(totalPages)">@totalPages</a></li>
+                    }
+
+                    <li class="page-item @if(currentPage == totalPages) { disabled }">
+                        <a class="page-link" href="@currentLink(checkMaxMin(currentPage + 1))" aria-label="Next">
+                            <span aria-hidden="true">&raquo;</span>
+                            <span class="sr-only">Next</span>
+                        </a>
+                    </li>
+                </ul>
+            </nav>
+        </div>
+    </div>
 }
