+++ conf/routes
@@ -9,9 +9,9 @@ GET     /*path/                                                     @controllers
 GET     /admin/user/:user                                           @controllers.Application.userAdmin(user)
 POST    /admin/user/:user/update                                    @controllers.Application.updateUser(user)
 GET     /admin/approval/versions                                    @controllers.Default.redirect(to = "/admin/approval/versions/pending")
-GET     /admin/approval/versions/pending                            @controllers.Application.showVersionQueuePending(page: Option[Int])
-GET     /admin/approval/versions/wip                                @controllers.Application.showVersionQueueWIP(page: Option[Int])
-GET     /admin/approval/versions/done                               @controllers.Application.showVersionQueueDone(page: Option[Int])
+GET     /admin/approval/versions/pending                            @controllers.Reviews.showVersionQueuePending(page: Option[Int])
+GET     /admin/approval/versions/wip                                @controllers.Reviews.showVersionQueueWIP(page: Option[Int])
+GET     /admin/approval/versions/done                               @controllers.Reviews.showVersionQueueDone(page: Option[Int])
 GET     /admin/approval/projects                                    @controllers.Application.showProjectVisibility()
 GET     /admin/flags                                                @controllers.Application.showFlags()
 GET     /admin/flags/:id/resolve/:resolved                          @controllers.Application.setFlagResolved(id: ObjectReference, resolved: Boolean)
